DTO management moved to the service logic
