+++ src/main/java/com/mateo9997/clubmanagementsystem/controller/PlayerController.java
@@ -30,8 +30,7 @@ public class PlayerController {
     public ResponseEntity<?> createPlayer(@PathVariable Long clubId, @RequestBody Player player) {
         verifyClubAccess(clubId);
         Player newPlayer = playerService.createPlayer(clubId, player);
-        PlayerDTO playerDTO = mapToDTO(newPlayer);
-        return ResponseEntity.ok(playerDTO);
+        return ResponseEntity.ok(newPlayer);
     }
 
     // List all players in a club
@@ -55,18 +54,16 @@ public class PlayerController {
     @GetMapping("/{playerId}")
     public ResponseEntity<PlayerDTO> getPlayerDetails(@PathVariable Long clubId, @PathVariable Long playerId) {
         verifyClubAccess(clubId);
-        Player player = playerService.getPlayerDetails(clubId, playerId);
-        PlayerDTO dto = mapToDTO(player);
-        return ResponseEntity.ok(dto);
+        PlayerDTO player = playerService.getPlayerDetails(clubId, playerId);
+        return ResponseEntity.ok(player);
     }
 
     // Update player details
     @PutMapping("/{playerId}")
     public ResponseEntity<PlayerDTO> updatePlayer(@PathVariable Long clubId, @PathVariable Long playerId, @RequestBody Player playerDetails) {
         verifyClubAccess(clubId);
-        Player updatedPlayer = playerService.updatePlayer(clubId, playerId, playerDetails);
-        PlayerDTO dto = mapToDTO(updatedPlayer);
-        return ResponseEntity.ok(dto);
+        PlayerDTO updatedPlayer = playerService.updatePlayer(clubId, playerId, playerDetails);
+        return ResponseEntity.ok(updatedPlayer);
     }
 
     // Delete a player
@@ -92,14 +89,4 @@ public class PlayerController {
         throw new IllegalStateException("Authenticated user is not of type ClubUserDetails");
     }
 
-    private PlayerDTO mapToDTO(Player player) {
-        PlayerDTO dto = new PlayerDTO();
-        dto.setId(player.getId());
-        dto.setGivenName(player.getGivenName());
-        dto.setFamilyName(player.getFamilyName());
-        dto.setNationality(player.getNationality());
-        dto.setEmail(player.getEmail());
-        dto.setDateOfBirth(player.getDateOfBirth());
-        return dto;
-    }
 }
