+++ src/main/java/com/mateo9997/clubmanagementsystem/service/PlayerService.java
@@ -1,5 +1,6 @@
 package com.mateo9997.clubmanagementsystem.service;
 
+import com.mateo9997.clubmanagementsystem.dto.PlayerDTO;
 import com.mateo9997.clubmanagementsystem.model.Club;
 import com.mateo9997.clubmanagementsystem.model.Player;
 import com.mateo9997.clubmanagementsystem.repository.ClubRepository;
@@ -36,7 +37,7 @@ public class PlayerService {
     }
 
     @Transactional
-    public Player updatePlayer(Long clubId, Long playerId, Player playerDetails) {
+    public PlayerDTO updatePlayer(Long clubId, Long playerId, Player playerDetails) {
         Player player = playerRepository.findByIdAndClubId(playerId, clubId)
                 .orElseThrow(() -> new NoSuchElementException("Player not found with id: " + playerId));
 
@@ -46,7 +47,8 @@ public class PlayerService {
         player.setEmail(playerDetails.getEmail());
         player.setDateOfBirth(playerDetails.getDateOfBirth());
 
-        return playerRepository.save(player);
+        playerRepository.save(player);
+        return mapToDTO(player);
     }
 
     @Transactional
@@ -58,8 +60,20 @@ public class PlayerService {
     }
 
     @Transactional
-    public Player getPlayerDetails(Long clubId, Long playerId) {
-        return playerRepository.findByIdAndClubId(playerId, clubId)
+    public PlayerDTO getPlayerDetails(Long clubId, Long playerId) {
+        Player player = playerRepository.findByIdAndClubId(playerId, clubId)
                 .orElseThrow(() -> new NoSuchElementException("Player not found with id: " + playerId + " and clubId: " + clubId));
+        return mapToDTO(player);
+    }
+
+    private PlayerDTO mapToDTO(Player player) {
+        PlayerDTO dto = new PlayerDTO();
+        dto.setId(player.getId());
+        dto.setGivenName(player.getGivenName());
+        dto.setFamilyName(player.getFamilyName());
+        dto.setNationality(player.getNationality());
+        dto.setEmail(player.getEmail());
+        dto.setDateOfBirth(player.getDateOfBirth());
+        return dto;
     }
 }
