+++ src/test/java/com/mateo9997/clubmanagementsystem/service/PlayerServiceTest.java
@@ -1,5 +1,6 @@
 package com.mateo9997.clubmanagementsystem.service;
 
+import com.mateo9997.clubmanagementsystem.dto.PlayerDTO;
 import com.mateo9997.clubmanagementsystem.model.Club;
 import com.mateo9997.clubmanagementsystem.model.Player;
 import com.mateo9997.clubmanagementsystem.repository.ClubRepository;
@@ -78,7 +79,7 @@ public class PlayerServiceTest {
         when(playerRepository.findByIdAndClubId(1L, 1L)).thenReturn(Optional.of(existingPlayer));
         when(playerRepository.save(existingPlayer)).thenReturn(existingPlayer);
 
-        Player updatedPlayer = playerService.updatePlayer(1L, 1L, newDetails);
+        PlayerDTO updatedPlayer = playerService.updatePlayer(1L, 1L, newDetails);
 
         assertEquals("Johnny", updatedPlayer.getGivenName());
         verify(playerRepository).save(existingPlayer);
@@ -117,7 +118,7 @@ public class PlayerServiceTest {
 
         when(playerRepository.findByIdAndClubId(1L, 1L)).thenReturn(Optional.of(player));
 
-        Player result = playerService.getPlayerDetails(1L, 1L);
+        PlayerDTO result = playerService.getPlayerDetails(1L, 1L);
 
         assertNotNull(result);
         assertEquals(1L, result.getId());
